add common method
