+++ src/test/resources/datasets/Invitation/getFindById-dataset.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <dataset>
-    <Invitations id="1" partnerId="3" code="Hello 1" dateEnd="2019-04-02" status="Ok"/>
+    <Invitations id="1" partnerId="3" code="Hello 1" dateEnd="2019-04-01" status="Ok"/>
 </dataset>
