+++ src/test/resources/project.properties
@@ -2,11 +2,11 @@ version=${project.version}
 artifactId=${project.artifactId}
 
 jdbc.driver=com.mysql.cj.jdbc.Driver
-db.url=jdbc:mysql://localhost:3306/timesheet_dev?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
+db.url=jdbc:mysql://localhost:3306/timesheet_dev?useSSL=false&useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
 db.username=root
-db.password=__chu2552
+db.password=qwerty
 schema=timesheet_dev
-table=project
+table=Invitations
 
 db.invitation-schema=timesheet_dev
 db.invitation-table-Invitations=Invitations
